Tela Inicial ✔

+++ app/Http/Controllers/DisciplinaController.php
@@ -5,25 +5,33 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Disciplina;
 use App\Models\Curso;
-use App\Models\Eixo;
+use App\Models\Docencia;
+use App\Facades\UserPermissions;
 
 use Illuminate\Support\Facades\Log;
 
 class DisciplinaController extends Controller
 {
 
+
     public function index()
     {
-
+        if(!UserPermissions::isAuthorized('disciplinas.index')) {
+            return response()->view('templates.restrito');
+        }
 
         $data = Disciplina::with(['curso'])
             ->orderBy('nome')->get();
-        
+        // return json_encode($data);
         return view('disciplinas.index', compact(['data']));
     }
 
     public function create()
     {
+
+        if(!UserPermissions::isAuthorized('disciplinas.create')) {
+            return response()->view('templates.restrito');
+        }
 
         $cursos = Curso::orderBy('nome')->get();
         return view('disciplinas.create', compact(['cursos']));
@@ -32,10 +40,10 @@ class DisciplinaController extends Controller
     public function validation(Request $request)
     {
 
-        $regras = [
+        $rules = [
             'nome' => 'required|max:100|min:10',
-            'carga' => 'required',
-            'curso' => 'required|max:12|min:1',
+            'carga' => 'required|max:12|min:1',
+            'curso' => 'required',
         ];
         $msgs = [
             "required" => "O preenchimento do campo [:attribute] é obrigatório!",
@@ -43,49 +51,48 @@ class DisciplinaController extends Controller
             "min" => "O campo [:attribute] possui tamanho mínimo de [:min] caracteres!",
         ];
 
-        $request->validate($regras, $msgs);
+        $request->validate($rules, $msgs);
     }
 
     public function store(Request $request)
     {
-        
 
         self::validation($request);
-
-        $total = Disciplina::where('nome', mb_strtoupper($request->nome, 'UTF-8'))
-            ->where('curso_id', $request->curso)
-            ->count();
-
-        
 
         $curso = Curso::find($request->curso);
         if (isset($curso)) {
             $obj = new Disciplina();
             $obj->nome = mb_strtoupper($request->nome, 'UTF-8');
             $obj->carga = $request->carga;
-            $obj->curso()->associate($curso); //ASSOCIAÇÃO DO NOVO OBJETO DE DISCIPLINA COM CURSO 
+            $obj->curso()->associate($curso);
             $obj->save();
+
             return redirect()->route('disciplinas.index');
         }
-        
-        return redirect()->route('disciplinas.index');
     }
 
     public function show($id)
     {
+        if(!UserPermissions::isAuthorized('disciplinas.show')) {
+            return response()->view('templates.restrito');
+        }
+        $doc = Docencia::with(['professor'])
+            ->where('disciplina_id', '=', $id)->distinct()->get(['professor_id']);
+
+        return view('disciplinas.show', compact(['doc']));
     }
 
     public function edit($id)
     {
+        if(!UserPermissions::isAuthorized('disciplinas.edit')) {
+            return response()->view('templates.restrito');
+        }
+
         $cursos = Curso::orderBy('nome')->get();
         $data = Disciplina::find($id);
 
         if (isset($data)) {
             return view('disciplinas.edit', compact(['data', 'cursos']));
-        } else {
-            $msg = "Disciplina";
-            $link = "disciplinas.index";
-            return view('erros.id', compact(['msg', 'link']));
         }
     }
 
@@ -94,37 +101,31 @@ class DisciplinaController extends Controller
 
         self::validation($request);
 
-        
 
         $curso = Curso::find($request->curso);
         $obj = Disciplina::find($id);
 
-        //PREENCHE OS CAMPOS COM OS DADOS DA DISCIPLINA SELECIONADA
         if (isset($obj) && isset($curso)) {
             $obj->nome = mb_strtoupper($request->nome, 'UTF-8');
             $obj->carga = $request->carga;
             $obj->curso()->associate($curso);
+
             $obj->save();
+
             return redirect()->route('disciplinas.index');
         }
-
-        
-        return redirect()->route('disciplinas.index');
     }
 
     public function destroy($id)
     {
-
+        if(!UserPermissions::isAuthorized('disciplinas.destroy')) {
+            return response()->view('templates.restrito');
+        }
         $obj = Disciplina::find($id);
 
         if (isset($obj)) {
             $obj->delete();
-        } else {
-            $msg = "Disciplina";
-            $link = "disciplinas.index";
-            return view('erros.id', compact(['msg', 'link']));
         }
-
         return redirect()->route('disciplinas.index');
     }
 }
